Move the default, domainType-derived properties to HalFormsPayloadMetadataContributor
For consistency, also have the properties derived from the domain type
itself be contributed instead of added directly inside the
HalFormsPayloadMetadataConverter.

This allows other HalFormsPayloadMetadataContributors to run either
before or after the default ones.
